cleaned standard header
