+++ src/IO/diff-checkpoint-posix.c
@@ -37,7 +37,6 @@
  */
 
 #include "interface.h"
-#include "macros.h"
 
 /*-------------------------------------------------------------------------*/
 /**
