+++ src/IO/ftiff.c
@@ -39,10 +39,10 @@
 #define _GNU_SOURCE
 
 #include "interface.h"
-#include "api_cuda.h"
-
-#include "utility.h"
-
+#include <sys/mman.h>
+#include <time.h>
+#include <sys/types.h>
+#include <dirent.h>
 /*  
 
     +-------------------------------------------------------------------------+
