+++ src/IO/posix.c
@@ -38,7 +38,6 @@
 
 
 #include "interface.h"
-#include "utility.h"
 
 /*-------------------------------------------------------------------------*/
 /**
