+++ src/macros.c
@@ -38,10 +38,6 @@
 
 
 
-#include <stdlib.h>
-#include <string.h>
-#include <stdio.h>
-#include <stdarg.h>
 #include "interface.h"
